requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
cachetools==5.3.1
pytest==7.4.0